refactor: Storage 를 ui 패키지로부터 분리
