+++ app/src/main/java/woowacourse/movie/data/storage/SettingsStorage.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.ui.storage
+package woowacourse.movie.data.storage
 
 import android.content.Context
 import android.content.Context.MODE_PRIVATE
