+++ app/src/main/java/woowacourse/movie/ui/fragment/settings/SettingsFragment.kt
@@ -7,7 +7,7 @@ import android.view.ViewGroup
 import androidx.appcompat.widget.SwitchCompat
 import androidx.fragment.app.Fragment
 import woowacourse.movie.R
-import woowacourse.movie.ui.storage.SettingsStorage
+import woowacourse.movie.data.storage.SettingsStorage
 
 class SettingsFragment : Fragment() {
